fix(test): correct model_dump expectation for WorkflowUUID
The test incorrectly expected model_dump() to serialize WorkflowUUID
to its short string format. However, Pydantic serialization only
applies in JSON mode (when_used=json). For model_dump(), the
object is returned as-is. Changed assertion to compare objects
directly.

+++ tests/registry/test_internal_workflow_router.py
@@ -92,8 +92,8 @@ class TestInternalWorkflowExecuteResponse:
             message="Workflow execution started",
         )
         data = resp.model_dump()
-        # WorkflowUUID serializes to its short() format
-        assert data["workflow_id"] == wf_id.short()
+        # model_dump() keeps WorkflowUUID as object (serialization only for JSON mode)
+        assert data["workflow_id"] == wf_id
         assert (
             data["workflow_execution_id"]
             == "wf-00000000000000000000000000000123:exec-456"
